Refactors the commander code and adds more E2E tests
